% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentclasslogit.R
\name{latentclasslogit}
\alias{latentclasslogit}
\title{Perform maximum-likelihood estimation for a latent class logistic regression model}
\usage{
latentclasslogit(formula, data, K = 2, start.beta = NULL,
  start.lambda = NULL, id = NULL, tol = 1e-05, theta.lower = NULL,
  theta.upper = NULL, method = "L-BFGS-B")
}
\arguments{
\item{formula}{a regression formula describing the relationship between the response and the covariates}

\item{data}{the data.frame containing the responses and covariates}

\item{K}{the number of mixtures (or latent classes)}

\item{start.beta}{a list of length K of starting values for each mixture's beta coefficients}

\item{start.lambda}{a vector of length K of starting values for the mixing proportions}

\item{tol}{a number specifying the tolerance used to determine convergence}

\item{theta.lower}{a numeric vector of lower bounds for the theta parameters}

\item{theta.upper}{a numeric vector of upper bounds for the theta parameters}

\item{method}{a string specifying the optimization routine to be used by optim}

\item{left}{a number specifying where left-censoring occurred}
}
\value{
a list containing the following elements:\cr
\item{beta}{a list containing the estimated regression coefficients}
\item{sigma}{a vector containing the estimated values of sigma}
\item{lambda}{a vector containing the estimated mixing proportions}
\item{delta}{a list of length K containing the estimated class membership probabilities for each observation}
\item{ll}{the log-likelihood function evaluated at the MLE}
}
\description{
This function performs maximum-likelihood estimation via the E-M algorithm to obtain
estimates of regression coefficients in a latent class logistic regression model.
}
