% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixturetobit.R
\name{mixturetobit}
\alias{mixturetobit}
\title{Perform maximum-likelihood estimation for a mixture of tobit regression models}
\usage{
mixturetobit(formula, data, K = 2, start.beta = NULL,
  start.sigma = NULL, start.lambda = NULL, id = NULL, left = -1,
  tol = 1e-05, theta.lower = NULL, theta.upper = NULL,
  method = "L-BFGS-B", beta.starting.sd = 0.5)
}
\arguments{
\item{formula}{a regression formula describing the relationship between the response and the covariates}

\item{data}{the data.frame containing the responses and covariates}

\item{K}{the number of mixtures (or latent classes)}

\item{start.beta}{a list of length K of starting values for each mixture's beta coefficients}

\item{start.sigma}{a vector of length K of starting values for each mixture's sigma value}

\item{start.lambda}{a vector of length K of starting values for the mixing proportions}

\item{id}{a string specifying the name of the column that identifies subjects}

\item{left}{a number specifying where left-censoring occurred}

\item{tol}{a vector of numbers specifying the tolerance(s) used to determine convergence. If length(tol) > 1, the methods used should be supplied as a vector to "method".}

\item{theta.lower}{a numeric vector of lower bounds for the theta parameters}

\item{theta.upper}{a numeric vector of upper bounds for the theta parameters}

\item{method}{a vector of strings specifying the optimization routine(s) to be used by optim. If length(method) > 1, optimization will be performed in succession with the tolerances provided to "tol"}

\item{beta.starting.sd}{a numeric for what std deviation to use for starting the EM with random betas}
}
\value{
a list containing the following elements:\cr
\item{beta}{a list containing the estimated regression coefficients}
\item{sigma}{a vector containing the estimated values of sigma}
\item{lambda}{a vector containing the estimated mixing proportions}
\item{delta}{a list of length K containing the estimated class membership probabilities for each observation}
\item{ll}{the log-likelihood function evaluated at the MLE}
}
\description{
This function performs maximum-likelihood estimation via the E-M algorithm to obtain
estimates of regression coefficients in a mixture of tobit regression models.
}
